You have a text defined in a String variable. Select one example sentence.
Create a method that finds the count of capital 'A' in a text. Use chartAt() method from String class.
Create a method that reverse the symbols in the text.  Use reverse() method from StringBuilder class.
Create a method that replaces 'A' with 'AA'. Use replaceAll() method from String class.
Create a method that splits the text in words. The method returns an array with all words in the sentence. Use split() method from String class.