Create a program that reads a text from a file and extracts all the unique words from that text.
The program takes a file as input argument and returns as result a collection of Strings containing unique words (no duplicated words).
The text is read from a file. File should contain multiple lines.
The text may contain words, spaces, and the following symbols .  , :  ;  ? !
To return the unique words choose a suitable collection in java. Which collection supports unique elements?