Create a simple Shopping Cart application. Implement a ShoppingCart class with the following methods:
addProduct (Product p, int amount )
addProduct (Product p ) – adds amount 1 of Product in shopping cart.
removeProduct(Product p, int amount )
removeProduct(Product p) - removes amount 1 of Product in shopping cart.
calculateTotalPrice() – calculates the total price of all products in the shopping cart.

Implement 2 different kind of discounts when calculating total price:
If the amount for a Product is more than 10, the price for that product should be discounted with 10%.
If the price of the whole shopping cart exceeds 1000, the total price should be discounted with 5%.