Create a class Person, with fields firstName, lastName, address and age. Override equals(Object obj) and hashCode() methods from java.lang.Object class. Override also toString method.
Create a main method to test the equality of objects. Follow the steps bellow:
Create a  new HashSet<Person> and try to add several Person objects to it. Add both: equal and non-equal objects.  Check that equal objects are not added twice. Try to understand how HashSet works, how it uses internally equals() and hashcode() methods to support unique elements.
Create a ArrayList<Person> and try to add some Person objects. Test the contains() method, which implementation is also based on equals() method.
Try to compare equal and non-equal people by using equals method, == operator and compareTo method. Test cases when two Person instances are the same and not the same compared with equals, == and compareTo. Try to understand how equals, == and comapreTo work.